refactor(livestream): separate internal viewer playback APIs from public platform APIs
- Add /api/internal/playback-info and /api/internal/sign-jwt, rate-limited, with no x402/BotID gates.
- Move shared playback-info and sign-jwt logic into lib/livepeer/*-handler.ts so both route variants reuse the same implementation.
- Restore /api/livepeer/playback-info and /api/livepeer/sign-jwt to their original gated behavior (x402 + BotID) for external/platform consumers.
- Point all first-party client callers to /api/internal/*:
  * useDetailPlaybackSources, useHeroPlaybackSource
  * WatchClient sign-jwt request
  * livepeer-thumbnails, lib/utils/thumbnail
  * build-lens-video-metadata
- Remove stale /api/livepeer/token-gate entry from instrumentation-client BotID protect list (route doesn't exist).

This makes the architecture explicit: first-party pages call internal APIs; external partners/payments use /api/livepeer/*.

Relates to #292, #291, #289.
